+++ crates/moon/tests/test_cases/test_filter/pkg_with_deps/lib/moon.pkg.json
