+++ crates/moon/tests/test_cases/test_filter/pkg_with_deps/main/moon.pkg.json
