+++ crates/moon/tests/test_cases/test_filter/pkg_with_test_imports/lib2/moon.pkg.json
