+++ crates/moon/tests/test_cases/test_filter/pkg_with_test_imports/lib3/moon.pkg.json
