+++ crates/moon/tests/test_cases/test_filter/pkg_with_test_imports/lib6/lib.mbt
